pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auto_test"
version = "0.1.0"
description = "API自动化测试项目"
authors = [{name = "Test Team"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pytest>=7.0.0",
    "requests>=2.28.0",
    "allure-pytest>=2.13.0",
    "loguru>=0.7.0",
    "faker>=18.0.0",
    "pyyaml>=6.0",
    "pytest-html>=3.1.0",
    "pytest-rerunfailures>=12.0",
    "pytest-timeout>=2.1.0",
    "pytest-xdist>=3.0.0",
    "pytest-metadata>=3.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--alluredir=./allure-results",
    "--clean-alluredir",
]
markers = [
    "smoke: 冒烟测试",
    "regression: 回归测试", 
    "api: API测试",
    "slow: 慢速测试",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
] 